refactor: controller와 view 메서드 재배치
